+++ Nagisa/MainPage.xaml.cpp
@@ -94,7 +94,8 @@ void Nagisa::MainPage::Button_Click(Platform::Object^ sender, Windows::UI::Xaml:
 			[this](IAsyncOperation<StorageFolder^>^ asyncInfo, AsyncStatus asyncStatus)
 	{
 		StorageFolder^ folder = asyncInfo->GetResults();
-		CoreApplication::MainView->CoreWindow->Dispatcher->RunAsync(
+
+		this->Dispatcher->RunAsync(
 			CoreDispatcherPriority::Normal, ref new DispatchedHandler([this, folder]()
 		{
 			if (folder)
@@ -114,7 +115,7 @@ void Nagisa::MainPage::Button_Click(Platform::Object^ sender, Windows::UI::Xaml:
 
 void Nagisa::MainPage::Page_Loaded(Platform::Object^ sender, Windows::UI::Xaml::RoutedEventArgs^ e)
 {
-	ConsoleWriteLine(L"M2-Team Nagisa Version 0.1.6");
+	ConsoleWriteLine(L"M2-Team Nagisa Version 0.1.7");
 	ConsoleWriteLine(L"© M2-Team. All rights reserved.");
 	ConsoleWriteLine(L"");