+++ NagisaCore/NagisaCore.BackgroundTask.cpp
@@ -1,0 +1,12 @@
+#include "pch.h"
+#include "NagisaCore.BackgroundTask.h"
+
+
+NagisaCore::BackgroundTask::BackgroundTask()
+{
+}
+
+void NagisaCore::BackgroundTask::Run(Windows::ApplicationModel::Background::IBackgroundTaskInstance ^ taskInstance)
+{
+	throw ref new Platform::NotImplementedException();
+}